+++ .gitignore
@@ -1,7 +1,9 @@
 #Doc des install
 listInstall.txt
 Documentation.txt
+commandePGSQL.txt
 exempleGenerate.sql
 gen2.sql
 .idea/*
 PyGestpgsql/backup/*
+PyGestpgsql/tmp/*
