--- PyGestpgsql/conf.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-#
-#Auteur Clément Cardarelli
-#Date Création 26/02
-#
-#
-#fichier contenant tous les paramêtres de l'application
-
-repertory = "/home/clement/Cours/B2/Base_de_Données/PostgreSQL/Projet/PostgreSQL/PyGestpgsql/"
-
-timer = "0 0 0 0 0"
